opus_agent_base/pyproject.toml
[project]
name = "opus-agent-base"
version = "0.1.0"
description = "Base framework for Opus AI Agents"
authors = [
    { name = "Sathish", email = "sathish316@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.21",
    "pydantic-ai>=0.8.1",
    "pydantic-ai-slim[anthropic,openai,retries]>=0.8.1",
    "tenacity>=8.2.0",
    "httpx>=0.24.0",
    "singleton-decorator>=1.0.0",
    "fastmcp>=2.12.4",
    "typer>=0.19.2",
    "tiktoken>=0.12.0",
    "chromadb>=1.1.1",
    "rapidfuzz>=3.14.1",
    "prompt-toolkit>=3.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
opus-agent-base = "opus_agent_base:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opus_agent_base"]
